code clean up

+++ Dog_bot/sound_direction.h
@@ -19,7 +19,7 @@ typedef enum {
 
 void sound_direction_setUp(void);
 
-int get_sound_angle(float* sound_direction);
+uint8_t get_sound_angle(float* sound_direction);
 
 
 #endif /* SOUND_DIRECTION_H */
